app/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.6
django-cors-headers==3.13.0
django-filter==22.1
django-redis-cache==3.0.1
djangorestframework==3.13.1
drf-yasg==1.20.0
environs==9.5.0
fcm-django==1.0.12
firebase-admin==5.2.0
flower==1.2.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
humanize==4.3.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.17.0
msgpack==1.0.4
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==3.5.3
requests==2.28.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.2
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
